2: Adds dependency injection sort of
I wanted to use Dagger for DI, but I couldn't wrap my head around it in less
than an hour, so I decided to just manually swap out dependencies as concrete
class instances.

LET IT BE KNOWN that DI is vastly superior to what I did in this commit, but for
a one-off code test, it gets the job done fine.