Adds error dialog to screen if fetching data fails
Lots of changes here.  Basically, the error statuses that we've been building
thus far are propagated upward to the Fragment layer, and the error dialog
reacts to the different types of error cases.

KNOWN ISSUE:  Since each fragment gets the data separately (which is really
inefficient because there are two HTTPS requests) only the active Peak fragment
will refresh when the user clicks "Retry".